+++ pandasdmx/api.py
@@ -14,7 +14,6 @@ import requests
 
 from . import remote
 from .reader import get_reader_for_content_type
-
 from .message import Message
 from pandasdmx import model
 from .model import DataStructureDefinition, MaintainableArtefact, ValidationLevels
@@ -61,8 +60,9 @@ class Request:
     #: :class:`.Session` for queries sent from the instance.
     session = None
 
-    def __init__(self, source=None, log_level=None, session=None,
-        timeout=30.1,     **session_opts):
+    def __init__(
+        self, source=None, log_level=None, session=None, timeout=30.1, **session_opts
+    ):
         """Constructor."""
         self.timeout = timeout
         try:
@@ -79,7 +79,6 @@ class Request:
         if log_level:
             logging.getLogger("pandasdmx").setLevel(log_level)
 
-
     @property
     def default_locale(self):
         """
@@ -89,11 +88,9 @@ class Request:
         """
         return model.DEFAULT_LOCALE
 
-
     @default_locale.setter
     def default_locale(self, locale):
         model.DEFAULT_LOCALE = locale
-
 
     @property
     def validation_level(self):
@@ -102,11 +99,9 @@ class Request:
         """
         return model.DEFAULT_VAL_LEVEL
 
-
     @validation_level.setter
     def validation_level(self, level):
         model.DEFAULT_VAL_LEVEL = ValidationLevels[level]
-
 
     def __getattr__(self, name):
         """Convenience methods."""
@@ -474,8 +469,7 @@ class Request:
             return req
 
         try:
-            response = self.session.send(req, 
-                timeout=self.timeout)
+            response = self.session.send(req, timeout=self.timeout)
             response.raise_for_status()
         except requests.exceptions.ConnectionError as e:
             raise e from None
@@ -492,7 +486,7 @@ class Request:
         response_content = remote.ResponseIO(response, tee=tofile)
 
         # Select reader class
-        content_type = response.headers.get("content-type", None)
+        content_type = response.headers.get("content-type")
         try:
             Reader = get_reader_for_content_type(content_type)
         except ValueError:
@@ -573,7 +567,84 @@ class Request:
             # No key is provided
             return list(all_keys)
 
+    def validate(self, msg, schema_dir=None):
+        """
+        Validate `msg` against the XML schemas which must
+        be installed first. 
+        
+        Parameters:
+        
+        msg: pandasdmx.message.Message or file-like
+            the XML message to be validated. If a
+            message.Message instance is provided, the file is
+            re-downloaded, ideally from cache.
+        schema_dir: path-like or str 
+            Optional custom dir where schemas
+            are installed. 
+            
+        Returns True on success.
+        
+        See also the LXML documentation.
+        """
+        # reload message file if message.Message object is provided rather than file-like
+        if isinstance(msg, Message): 
+            msg_response = self.session.get(msg.response.url)
+            msg = remote.ResponseIO(msg_response)
+        # Select reader class
+        # Currently, the only reader that can validate
+        # sdmx files is the sdmxml reader.
+        from .reader.sdmxml import Reader
+
+        # Validate message against the schema referenced in the msg
+        return Reader.validate_message(msg, schema_dir=schema_dir)
+
 
 def read_url(url, **kwargs):
     """Request a URL directly."""
     return Request().get(url=url, **kwargs)
+
+
+def install_schemas(schema_dir=None, **kwargs):
+    """
+    Download the complete set of XML schemas from `sdmx.org <http://www.sdmx.org>`_. 
+    and install them in <schema_dir>. The schemas
+    are included in Section 3b of the SDMXML 2.1 standard. Installation
+    steps are logged.
+    
+    Parameters:
+    
+    schema_dir path-like or str 
+        defaults to the platform-specific appdata dir 
+        of the user. <appname> 
+        is set to "pandasdmx".
+    **kwargs: 
+        optional kwargs passed to
+        `requests.get()` to configure the 
+        HTTP connection, eg. `verify`or `proxies`.
+        
+    Returns None on success.
+    """
+    from zipfile import ZipFile
+    from .reader.sdmxml import Reader
+    from pathlib import Path
+    import os
+
+    url = "https://sdmx.org/wp-content/uploads/SDMX_2-1_SECTION_3B_SDMX_ML_Schemas_Samples_2020-07.zip"
+    logger.info("Downloading SDMX 2.1 Standard, Section 3b from www.sdmx.org...")
+    response = requests.get(url=url, **kwargs)
+    content = remote.ResponseIO(response)
+    zf = ZipFile(content.tee)
+    logger.info("Done.")
+    schema_dir = Path(schema_dir or Reader.get_schema_dir())
+    # Create any non-existent dirs
+    os.makedirs(os.path.dirname(filepath), exist_ok=True)    
+    logger.info(f"Installing schema files in {str(schema_dir)}...")
+    for s in zf.infolist():
+        if s.filename.startswith("schemas/"):
+            with zf.open(s, "r") as src:
+                # cut off the previx "schemas/"
+                fn = s.filename[8:]
+                filepath = schema_dir.joinpath(fn)
+                with open(filepath, "wb") as target:
+                    target.write(src.read())
+    logger.info("Done.")