+++ pandasdmx/reader/base.py
@@ -1,3 +1,10 @@
+from pathlib import Path
+
+try:
+    import appdirs
+except ImportError:
+    pass
+
 import logging
 from abc import ABC, abstractmethod
 from functools import lru_cache
@@ -62,3 +69,11 @@ class BaseReader(ABC):
             An instance of a Message subclass.
         """
         pass  # pragma: no cover
+
+    @staticmethod
+    def get_schema_dir():
+        return Path(appdirs.user_data_dir(appname="pandasdmx", appauthor=False))
+
+    @staticmethod
+    def validate_message(msg, schema_dir=None):
+        raise NotImplementedError